requirements.txt
Flask
Jinja2
Werkzeug
itsdangerous
MarkupSafe
python-dotenv
flask_sqlalchemy
Flask-Login
Flask-WTF
WTForms
email-validator
Flask-Bootstrap
Flask-Mail
SQLAlchemy
PyMySQL
folium
APScheduler
pycparser
gunicorn
huggingface_hub